requirements.txt
astroid==2.5.2
astunparse==1.6.3
async-generator==1.10
cryptography==2.8
fasteners==0.14.1
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
gast==0.3.3
h5py==2.10.0
imageio==2.9.0
imageio-ffmpeg==0.4.3
imutils==0.5.4
jedi==0.18.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.4.1
matplotlib-inline==0.1.2
munkres==1.1.4
numpy==1.21.0
opencv-python==4.5.1.48
openvino-ubuntu20==2021.2
opt-einsum==3.3.0
pandas==1.2.3
Pillow==8.1.1
protobuf==3.15.7
pylint==2.7.4
pylint-protobuf==0.20.0
pypolaris==0.0.2
PyYAML==5.4.1
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.1
sklearn==0.0
streamlit==0.83.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==1.0.1
yolov4==2.0.2
